requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
pycryptodome==3.19.0
requests==2.31.0
